Tranladar funciones de propietario a cabezera
