+++ menu.h
@@ -3,12 +3,10 @@
 
 void Menu();
 void menuPropietario();
-void RegistrarPropietario();
-void listaDePropietarios();
+
 void menuInventario();
 void RegistrarProducto();
-void eliminarPropietario();
-void actualizarPropietario();
+
 void RegistrarMascota();
 void menuMascotas();
 void mostrarInventario();
