+++ propietario.h
@@ -1,0 +1,10 @@
+#ifndef PROPIETARIO_H
+#define PROPIETARIO_H
+
+void RegistrarPropietario();
+void listaDePropietarios();
+void eliminarPropietario();
+void actualizarPropietario();
+
+#endif
+
